adk-agent-dev/requirements.txt
pandas==2.0.0
numpy==1.24.3
influxdb-client==1.36.0
python-dotenv==1.0.0
ccxt==4.4.77
requests==2.31.0
python-dateutil==2.8.2
websocket-client==1.6.1